requirements.txt
jupyterlab
pandas
scikit-learn
matplotlib
seaborn
openai
requests
feedparser
pdfplumber
mysql-connector-python
